+++ 1.1练习作业提交/maizhenchao/1.1homework/homework.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+<head lang="en">
+    <meta charset="UTF-8">
+    <title>The blah</title>
+    <link rel="stylesheet" type="text/css" href="homework.css">
+</head>
+<body>
+    <div class="header">
+        <img src="images/blah.png">
+        <ul class="nav">
+            <li><a href="#">Home</a> </li>
+            <li><a href="#">Site</a> </li>
+            <li><a href="#">Other</a> </li>
+        </ul>
+    </div>
+    <div class="main-content">
+        <h2>The Beach</h2>
+        <hr>
+        <ul class="photos">
+            <li><img src="images/0001.jpg" width="150" height="150"></li>
+            <li><img src="images/0002.jpg" width="150" height="150"></li>
+            <li><img src="images/0003.jpg" width="150" height="150"></li>
+        </ul>
+        <p>Say Something</p>
+    </div>
+    <div class="footer">
+        <p>&copy;mugglecoding</p>
+    </div>
+
+</body>
+</html>
